+++ python/labours/__main__.py
@@ -1,7 +1,6 @@
 import sys
 
-from labours.labours import main
-
+from labours.cli import main
 
 if __name__ == "__main__":
     sys.exit(main())
