+++ python/labours/_vendor/swivel.py
@@ -59,6 +59,7 @@ import threading
 import time
 
 import numpy as np
+
 import tensorflow as tf
 from tensorflow.python.client import device_lib
