+++ python/setup.cfg
@@ -4,6 +4,7 @@ ignore = D,B007
 import-order-style = appnexus
 inline-quotes = "
 max-line-length = 99
+application-import-names = labours
 
 [isort]
 force_sort_within_sections = true
